requirements.txt
rawpy==0.23.1
opencv-python==4.8.0.74
numpy==1.26.4
pyexiftool==0.5.5
scikit-learn==1.3.1
tqdm==4.65.0
omegaconf==2.3.0
torchinfo==1.8.0
lightning==2.1.2
hydra-core==1.3.2
torchmetrics==1.0.1
rich==13.8.1
aim==3.17.5
imageio==2.35.1
hydra-slurm-rich-launcher==1.1.0